requirements.txt
opencv-python
numpy
baidu-aip